pyproject.toml
[project]
name = "arvo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Dan Greco", email = "git@dangre.co" }]
requires-python = ">=3.13"
dependencies = [
  "argon2-cffi>=25.1.0",
  "flask>=3.1.2",
  "flask-cors>=6.0.2",
  "flask-pydantic>=0.14.0",
  "flask-sqlalchemy>=3.1.1",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.12.0",
  "pyjwt>=2.10.1",
  "sqlalchemy>=2.0.45",
  "werkzeug>=3.1.4",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.4"]

[project.scripts]
arvo = "arvo:main"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[tool.ty.environment]
python = ".venv"
